requirements.txt
openai>=1.0.0  
tiktoken  
qdrant-client  
langchain  
langchain-community  
python-dotenv  
